blockbuilder/scheduler: encapsulate partitionState offset access
Route the scheduler's direct reads of the committed and planned
advancingOffsets through methods on partitionState, add a
newPartitionState constructor, and move advancingOffset alongside
partitionState since it is its only user.
