+++ pkg/blockbuilder/scheduler/partition_state.go
@@ -7,6 +7,9 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/go-kit/log"
+	"github.com/go-kit/log/level"
+
 	"github.com/grafana/mimir/pkg/blockbuilder/schedulerpb"
 )
 
@@ -39,6 +42,28 @@ const (
 	bucketSame   = 0
 	bucketAfter  = 1
 )
+
+func newPartitionState(topic string, partition int32, metrics *schedulerMetrics, logger log.Logger) *partitionState {
+	return &partitionState{
+		topic:          topic,
+		partition:      partition,
+		pendingJobs:    list.New(),
+		plannedJobs:    list.New(),
+		plannedJobsMap: make(map[string]*jobState),
+		planned: &advancingOffset{
+			name:    offsetNamePlanned,
+			off:     offsetEmpty,
+			metrics: metrics,
+			logger:  logger,
+		},
+		committed: &advancingOffset{
+			name:    offsetNameCommitted,
+			off:     offsetEmpty,
+			metrics: metrics,
+			logger:  logger,
+		},
+	}
+}
 
 // updateEndOffset processes an end offset and returns a consumption job spec if
 // one is ready. This is expected to be called with monotonically increasing
@@ -126,4 +151,95 @@ func (s *partitionState) completeJob(jobID string) error {
 		s.committed.advance(js.jobID, js.spec)
 	}
 	return nil
+}
+
+func (s *partitionState) committedOffset() int64 {
+	return s.committed.offset()
+}
+
+func (s *partitionState) committedEmpty() bool {
+	return s.committed.empty()
+}
+
+func (s *partitionState) committedBeyondSpec(spec schedulerpb.JobSpec) bool {
+	return s.committed.beyondSpec(spec)
+}
+
+func (s *partitionState) plannedOffset() int64 {
+	return s.planned.offset()
+}
+
+func (s *partitionState) plannedEmpty() bool {
+	return s.planned.empty()
+}
+
+func (s *partitionState) plannedBeyondSpec(spec schedulerpb.JobSpec) bool {
+	return s.planned.beyondSpec(spec)
+}
+
+func (s *partitionState) plannedValidNextSpec(spec schedulerpb.JobSpec) bool {
+	return s.planned.validNextSpec(spec)
+}
+
+// advancingOffset keeps track of an offset that is expected to advance
+// monotonically based on job progression.
+type advancingOffset struct {
+	off     int64
+	name    string
+	metrics *schedulerMetrics
+	logger  log.Logger
+}
+
+const offsetEmpty int64 = -1
+
+const (
+	offsetNamePlanned   = "planned"
+	offsetNameCommitted = "committed"
+)
+
+// advance moves the offset forward by the given job spec. Advancements are
+// expected to be monotonically increasing and contiguous. Advance will not
+// allow backwards movement. If a gap is detected, a warning is logged and a
+// metric is incremented.
+func (o *advancingOffset) advance(jobID string, spec schedulerpb.JobSpec) {
+	if o.beyondSpec(spec) {
+		// Frequent, and expected.
+		level.Debug(o.logger).Log("msg", "ignoring historical job", "offset_name", o.name, "job_id", jobID,
+			"partition", spec.Partition, "start_offset", spec.StartOffset, "end_offset", spec.EndOffset, "committed", o.off)
+		return
+	}
+
+	if !o.validNextSpec(spec) {
+		// Gap detected.
+		level.Warn(o.logger).Log("msg", "gap detected in offset advancement", "offset_name", o.name, "job_id", jobID,
+			"partition", spec.Partition, "start_offset", spec.StartOffset, "end_offset", spec.EndOffset, "committed", o.off)
+		o.metrics.jobGapDetected.WithLabelValues(o.name).Inc()
+	}
+
+	o.off = spec.EndOffset
+}
+
+func (o *advancingOffset) offset() int64 {
+	return o.off
+}
+
+func (o *advancingOffset) set(offset int64) {
+	o.off = offset
+}
+
+// empty returns true if the offset is empty and uninitialized.
+func (o *advancingOffset) empty() bool {
+	return o.off == offsetEmpty
+}
+
+// validNextSpec returns true if the given job spec is valid to be added to the
+// offset. It is valid if the start offset is the same as the current offset.
+// We also allow transitioning out of an empty offset without calling it a gap.
+func (o *advancingOffset) validNextSpec(spec schedulerpb.JobSpec) bool {
+	return o.off == spec.StartOffset || o.empty()
+}
+
+// beyondSpec returns true if the offset is beyond the given job spec.
+func (o *advancingOffset) beyondSpec(spec schedulerpb.JobSpec) bool {
+	return !o.empty() && spec.EndOffset <= o.off
 }
